fastapi-course-1/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
fastapi==0.115.12
gunicorn==23.0.0
h11==0.16.0
idna==3.10
packaging==25.0
pydantic==2.11.4
pydantic_core==2.33.2
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlmodel==0.0.24
starlette==0.46.2
timescaledb
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
python-decouple
psycopg[binary]
